Move header under components
