+++ client/gutenboarding/components/header/style.scss
@@ -1,4 +1,4 @@
-@import '../gutenberg-styles/styles';
+@import '../../gutenberg-styles/styles';
 
 // Copied from https://github.com/WordPress/gutenberg/blob/8c0c7b7267473b5c197dd3052e9707f75f4e6448/packages/edit-widgets/src/components/header/style.scss
 .gutenboarding__header {
